Add files via upload
Change the date selection from Oct-2020 to Nov-2020. Also add a try/except to check if Nov-2020 is gone, and if so, pick Dec-2020. Exception will rise when current month is past over time. 
Also remove code for the TO date as website has set the default TO date to current month.